+++ changelog.markdown
@@ -122,4 +122,4 @@
 * Added module **ToolShed.Test.Ix** to facilitate testing instances of class **Data.Array.IArray.Ix**.
 * Added module **ToolShed.Data.Array.IArray**.
 * Added function **ToolShed.Data.String.deTab**.
-* Generalised function **ToolShed.System.Random.select** to accept any **Data.Foldable.Foldable** rather than merely a list.
+* Generalised function **ToolShed.System.Random.select** to accept any **Data.Foldable.Foldable** rather than merely a list, & to return **Maybe** rather than an error on receipt of null.